+++ client-roomfinder/package.json
@@ -20,6 +20,7 @@
     "axios": "^1.6.8",
     "cloudinary": "^2.1.0",
     "framer-motion": "^11.0.24",
+    "javascript-time-ago": "^2.5.10",
     "leaflet": "^1.9.4",
     "leaflet-defaulticon-compatibility": "^0.1.2",
     "next": "14.1.4",
