--- client-roomfinder/src/app/message/page.tsx
@@ -1,67 +1,0 @@
-'use client';
-import { useEffect, useState } from "react";
-import io from "socket.io-client";
-
-const socket = io("http://localhost:1234");
-
-interface Message {
-  body: string;
-  from: string;
-}
-
-export default function MessageComponent() {
-  const [messages, setMessages] = useState<Message[]>([]);
-  const [message, setMessage] = useState("");
-
-  const receiveMessage = (message: Message) => {
-    setMessages((prevMessages) => [...prevMessages, message]);
-  }
-
-  useEffect(() => {
-    socket.on("message", receiveMessage)
-
-    return () => {
-      socket.off("message", receiveMessage);
-    };
-  }, []);
-
-  const handleSubmit = (event: React.FormEvent) => {
-    event.preventDefault();
-    const newMessage = {
-      body: message,
-      from: "You",
-    };
-    setMessages(state => [newMessage, ...state]);
-    setMessage("");
-    socket.emit("message", newMessage.body);
-  };
-
-  return (
-    <div className="h-screen bg-zinc-800 text-white">
-      <form onSubmit={handleSubmit} className="bg-zinc-900 p-10">
-        <h1 className="text-2xl font-bold my-2">Chat Next</h1>
-        <input
-          name="message"
-          type="text"
-          placeholder="Write your message..."
-          onChange={(e) => setMessage(e.target.value)}
-          className="border-2 border-zinc-500 p-2 w-full text-black"
-          value={message}
-          autoFocus
-        />
-
-        <ul className="h-80 overflow-y-auto">
-          {messages.map((message, index) => (
-            <li
-              key={index}
-              className={`my-2 p-2 table text-sm rounded-md ${message.from === "You" ? "bg-sky-700 ml-auto" : "bg-black"
-                }`}
-            >
-              <b>{message.from}</b>:{message.body}
-            </li>
-          ))}
-        </ul>
-      </form>
-    </div>
-  );
-}
